@@ -10,28 +10,35 @@
         "@emotion/styled": "^11.10.5",
         "@material-ui/core": "^4.12.4",
         "@mui/icons-material": "^5.11.0",
-        "@mui/material": "^5.11.4",
+        "@mui/material": "^5.11.8",
         "and": "^0.0.3",
-        "axios": "^1.2.2",
+        "axios": "^1.3.2",
+        "bootstrap-css-only": "^4.4.1",
+        "fontawesome-free": "^1.0.4",
+        "mdb-ui-kit": "^6.1.0",
+        "mdbreact": "^4.6.1",
         "node.js": "^0.0.1-security",
         "react-icons": "^4.7.1",
-        "react-router-dom": "^6.6.2",
+        "react-router-dom": "^6.8.1",
         "react-scripts": "^5.0.1"
     },
     "devDependencies": {
         "@babel/core": "^7.20.12",
         "@babel/preset-env": "^7.20.2",
         "@babel/preset-react": "^7.18.6",
+        "@testing-library/jest-dom": "^5.16.5",
         "@testing-library/react": "^12.1.5",
-        "@typescript-eslint/eslint-plugin": "^5.51.0",
-        "@typescript-eslint/parser": "^5.51.0",
+        "@typescript-eslint/eslint-plugin": "^5.54.0",
+        "@typescript-eslint/parser": "^5.54.0",
+        "babel-loader": "^9.1.2",
         "css-loader": "^6.7.3",
-        "eslint": "^8.33.0",
+        "eslint": "^8.35.0",
         "eslint-plugin-react": "^7.32.2",
         "file-loader": "^6.2.0",
         "react": "^17.0.2",
         "react-dom": "^17.0.2",
         "style-loader": "^3.3.1",
+        "webpack": "^5.75.0",
         "webpack-cli": "^5.0.1"
     }
 }
