@@ -1,9 +1,12 @@
 from django.urls import path
-from .views import UserView, CalculatorView, LoginView
- 
+from .views import UserView, CalculatorView, LoginView, CustomAuthToken, CalculatorConstantsView
+from rest_framework.authtoken import views
  
 urlpatterns = [
     path('user/', UserView.as_view()),
     path('calculator/', CalculatorView.as_view()),
     path('login/', LoginView.as_view()),
+    path('calculatorconstants/', CalculatorConstantsView.as_view()),
+    path('api-token-auth/', CustomAuthToken.as_view())
 ]
+
