@@ -1,6 +1,7 @@
 import React, { Component } from "react";
 import Landing from "../../static/images/tree.jpeg";
 import FoodAtPub from "../../static/images/food_at_pub.png";
+import Roadmap from "../../static/images/roadmap.png";
 import Calculator from "../../static/images/calculator_image.jpg";
 import "../../static/css/base.css";
 import "../../static/css/home.css";
@@ -24,8 +25,8 @@ export default class HomePage extends Component {
                 
                 <div className="text_div">
                     <div id="intro_text">
-                        <h2 className="text_heading">Introduction</h2>
-                        <p className="text">This programme supports our clients to cut
+                        <h2 className="text_heading">Planet First Locals</h2>
+                        <p className="text">This programme supports our clients
                         in cutting their carbon, by understanding their business emissions, including 
                         their supply chains, and setting and achieving ambitious carbon reduction 
                         targets.<br></br><br></br>
@@ -45,23 +46,23 @@ export default class HomePage extends Component {
 
                 <div className="text_div">
                     <div id="other_text">
-                        <h2 className="text_heading">Other Heading</h2>
-                        <p className="text">Animi et voluptatem qui sed dolores velit aperiam et. Laborum est corrupti aut sit excepturi qui modi. Illo qui illum ut molestiae quam ut. In dolore esse tempore et possimus vitae ea. Assumenda similique quaerat velit. Ipsam aut laudantium ullam quis enim quae dolorem placeat.</p>
+                        <h2 className="text_heading" id="heading_right_adjusted">Take the Pledge</h2>
+                        <p className="text">With support from Small World Consulting, we will estimate your Scope 1, 2 and 3 footprint (to learn more about Scope 1, 2 and 3 see next tab)</p>
+                        <p className="text">Take the Planet First Locals pledge to reduce emissions by at least 15% of the next 12 months (in line with 1.5°C Paris target)</p>
+                        <p className="text">The entire carbon footprint remaining is then removed by Brewdog through Lost Forest and Dark Forest projects</p>
+                    
                     </div>
-                    <img className="text_pic" id="pic_other" src={FoodAtPub} alt="Picture of some food at a Brewdog pub"/>
+                    <img className="text_pic" id="pic_other" src={Roadmap} alt="Picture of an ecological roadmap"/>
                 </div>
 
                 <div className="calculator_div">
                     <div id="calculator_text">
                         <h2 className="text_heading">Footprint Calculator</h2>
-                        <p className="text">This programme supports our clients to cut
-                        in cutting their carbon, by understanding their business emissions, including 
-                        their supply chains, and setting and achieving ambitious carbon reduction 
-                        targets.<br></br><br></br>
-                        Our aim is to build a network of dynamic and 
-                        proactive bar and restaurant innovators that are 
-                        committed to going as low carbon as they can, and 
-                        then some.<br></br><br></br>
+                        <p className="text">This programme helps our clients to reduce their carbon emissions,
+                        by understanding their business emissions and supply chain emissions. <br></br><br></br>
+                        Our aim is to build a network of dynamic and proactive bar and restaurant innovators that 
+                        are committed to going as low carbon as they can, and then some.
+                        <br></br><br></br>
                         Real innovation happens when individuals get 
                         inspired to look beyond business as usual and 
                         commit to doing things differently. Here at BrewDog, 
@@ -69,11 +70,12 @@ export default class HomePage extends Component {
                         is future-ready and supportive of those who are ready to 
                         join us.</p>
                         <form action="/carboncalculator">
-                            <button type="submit" value="&#9654; Launch">&#9654; Launch</button>
+                            <button type="submit" data-testid="launch_button" value="&#9654; Launch">&#9654; Get Started</button>
                         </form>
                     </div>
                     <img className="text_pic" id="pic_calculator" src={Calculator} alt="Picture of a person using a calculator"/>
                 </div>
+                <br /><br />
 
 
             </div>
