@@ -1,6 +1,9 @@
 import React from "react";
 import { useNavigate } from "react-router-dom";
-import { Stack} from "@mui/material";
+
+
+
+import "../../static/css/useraccount.css";
 
 const SignUp = () => {
     const navigate = useNavigate();
@@ -8,52 +11,76 @@ const SignUp = () => {
     const handleSubmit = (event) => {
         event.preventDefault();
         const data = new FormData(event.target);
+
+        if (data.get("password") !== data.get("confirm-password")) {
+            window.alert("Passwords do not match");
+            return;
+        }
+
         fetch("/brewdog/user/", {
             method: "POST",
             body: data,
             credentials: "include"
         }).then(response => {
             if (response.ok) {
-                navigate("/carboncalculator");
+                navigate("/login");
             } else {
-                console.log("Error");
+                response.text().then( text => {{ 
+                    window.alert(text);
+                    throw new Error(text); }});
             }
+        }).catch( error => {
+            console.log("Error: " + error );
         });
     };
 
+
     return (
-        <div>
-            <div>
-            <h1>Create Account</h1>
-            <form method="POST" credentials="include" onSubmit={handleSubmit}>
-                <input type="hidden" name="csrfmiddlewaretoken" value="csrftoken"/>
-                <Stack spacing={3} direction="column">
-                <label>
-                    Name:
-                    <input type="text" name="name" />
-                </label>
-                <label>
-                    Email:
-                    <input type="text" name="email" />
-                </label>
-                <label>
-                    Company:
-                    <input type="text" name="company" />
-                </label>
-                <label>
-                    Phone:
-                    <input type="text" name="phone" />
-                </label>
-                <label>
-                    Password:
-                    <input type="text" name="password" />
-                </label>
-                <input type="submit" value="Submit" />
-                </Stack>
-            </form>
+        <>
+            <div className="flex-container">
+                
+                <form className="account_form" method="POST" credentials="include" onSubmit={handleSubmit}>
+                    <input type="hidden" name="csrfmiddlewaretoken" value="csrftoken"/>
+                    <h2>Create Account</h2>
+                    <div className="form-outline mb-2 field_container">
+                        <label className="form-label form-input-label" htmlFor="signup-form-username">Username</label>
+                        <input type="text" name="username" id="signup-form-username" className="form-control form-input-field" required/>
+                    </div>
+
+                    <div className="form-outline mb-2 field_container">
+                        <label className="form-label form-input-label" htmlFor="signup-form-email">Email address</label>
+                        <input type="email" name="email" id="signup-form-email" className="form-control form-input-field" required/>
+                    </div>
+
+                    <div className="form-outline mb-2 field_container">
+                        <label className="form-label form-input-label" htmlFor="signup-form-company">Company</label>
+                        <input type="text" name="company" id="signup-form-company" className="form-control form-input-field" required/>
+                    </div>
+
+                    <div className="form-outline mb-2 field_container">
+                        <label className="form-label form-input-label" htmlFor="signup-form-phone-number">Phone number</label>
+                        <input type="text" name="phone" id="signup-form-phone" className="form-control form-input-field" required/>
+                    </div>
+
+                    <div className="form-outline mb-2 field_container">
+                        <label className="form-label form-input-label" htmlFor="signup-form-password">Password</label>
+                        <input type="password" name="password" id="signup-form-password" className="form-control form-input-field" required/>
+                    </div>
+
+                    <div className="form-outline mb-2 field_container">
+                        <label className="form-label form-input-label" htmlFor="signup-form-confirm-password">Confirm password</label>
+                        <input type="password" name="confirm-password" id="signup-form-confirm-password" className="form-control form-input-field" required/>
+                    </div>
+
+
+                    <button type="submit" className="btn btn-primary btn-block ripple-effect">Submit</button>
+                    <p>Already have an account?<br/>Login <a href="/login"><strong>here</strong></a>.</p>
+
+                </form>
             </div>
-        </div>
-        );
+        </>
+    );
 };
+
 export default SignUp;
 
